+++ app/controllers/secrets_controller.rb
@@ -2,21 +2,21 @@ class SecretsController < ApplicationController
   before_action :authenticate_request
 
   def create
-    req = CertIssueRequest.new(params_permitted)
+    req = Requests::CreateSecretRequest.new(params_permitted)
     if !req.valid?
       raise BadRequestError.new req.errors.full_messages
     end
-    result = IssueCert.call(request: req, identity: identity)
+    result = CreateSecret.call(request: req, identity: identity)
     if result.failure?
       raise (result.error || StandardError.new(result.message))
     end
-    @cert = result.cert
+    @secret = result.secret
   end
 
   private
 
   def params_permitted
-    attrs = CertIssueRequest.new.attributes.keys
+    attrs = Requests::CreateSecretRequest.attributes.keys
     params.require(:secret_request).permit(attrs)
   end
 end
