+++ app/lib/requests/create_secret_request.rb
@@ -1,0 +1,11 @@
+module Requests
+  class CreateSecretRequest
+    include ActiveModel::Model
+    include ActiveModel::Attributes
+
+    attribute :path, :string
+    attribute :data, :hash
+
+    validates :path, presence: true
+  end
+end
