+++ test/interactors/refresh_domain_test.rb
@@ -4,7 +4,7 @@ class RefreshDomainTest < ActiveSupport::TestCase
   def setup
     @domain = domains(:owner_match)
     @identity = Identity.new(subject: @domain.users_array.first)
-    @cr = CertIssueRequest.new(common_name: @domain.fqdn)
+    @cr = Requests::CertIssueRequest.new(common_name: @domain.fqdn)
     @interactor = RefreshDomain
   end
 
@@ -19,7 +19,7 @@ class RefreshDomainTest < ActiveSupport::TestCase
 
   test ".call deletes db record when service 404" do
     @domain = domains(:no_match) # this fixture should have no match
-    @cr = CertIssueRequest.new(common_name: @domain.fqdn)
+    @cr = Requests::CertIssueRequest.new(common_name: @domain.fqdn)
     rslt = @interactor.call(identity: @identity, request: @cr)
     assert rslt.success?
     reloaded = Domain.where(fqdn: @domain.fqdn).first
